Response\HeadersTest: split `getIterator()` specific tests to dedicated test class
